clang-tidy: Enable bugprone-narrowing-conversions check
This check indicates instances where there's a potential loss of
data during type conversions, particularly when converting from
a wider type to a narrower type.

+++ sync_watch.cpp
@@ -114,7 +114,7 @@ int SyncWatch::callback(sd_event_source* /* s */, int fd, uint32_t revents,
         }
 
         // fileMap<wd, path>
-        auto rc = syncWatch->syncCallback(event->mask,
+        auto rc = syncWatch->syncCallback(static_cast<int>(event->mask),
                                           syncWatch->fileMap[event->wd]);
         if (rc)
         {
